store/tikv:adddress comments
